so i got some stuff working, but i think its time to shift to a different data model because this is very unreliable and i get exploding gradients

+++ optimizer.h
@@ -5,9 +5,9 @@
 typedef struct {
   Variable* weights;
   Variable* biases;
-  float learning_rate;
+  double learning_rate;
 } SGDOptimizer;
 
-SGDOptimizer optimizer_sgd_create(Variable* weights, Variable* biases, float learning_rate);
+SGDOptimizer optimizer_sgd_create(Variable* weights, Variable* biases, double learning_rate);
 void optimizer_sgd_step(SGDOptimizer* optimizer);
 void optimizer_sgd_zero_grad(SGDOptimizer* optimizer);
